+++ src/manager_inst.h
@@ -47,7 +47,10 @@ class ManagerInst
   void WriteProducerInformation(cyclus::toolkit::CommodityProducer*
                                 producer);
 
- private:
+  private:
+  // Code Injection:
+  #include "toolkit/position.cycpp.h"
+
   /// register a child
   void Register_(cyclus::Agent* agent);
 
@@ -65,27 +68,6 @@ class ManagerInst
   }
   std::vector<std::string> prototypes;
 
-  #pragma cyclus var { \
-    "default": 0.0, \
-    "uilabel": "Geographical latitude in degrees as a double", \
-    "doc": "Latitude of the agent's geographical position. The value " \
-           "should be expressed in degrees as a double." \
-  }
-  double latitude;
-
-  #pragma cyclus var { \
-    "default": 0.0, \
-    "uilabel": "Geographical longitude in degrees as a double", \
-    "doc": "Longitude of the agent's geographical position. The value " \
-           "should be expressed in degrees as a double." \
-  }
-  double longitude;
-  // clang-format on
-
-  cyclus::toolkit::Position coordinates;
-
-  /// Records an agent's latitude and longitude to the output db
-  void RecordPosition();
 };
 
 }  // namespace cycamore