+++ src/mixer.h
@@ -128,29 +128,10 @@ class Mixer
   //// A policy for sending material
   cyclus::toolkit::MatlSellPolicy sell_policy;
 
- private:
-  // clang-format off
-  #pragma cyclus var { \
-    "default": 0.0, \
-    "uilabel": "Geographical latitude in degrees as a double", \
-    "doc": "Latitude of the agent's geographical position. The value should " \
-           "be expressed in degrees as a double." \
-  }
-  double latitude;
-
-  #pragma cyclus var { \
-    "default": 0.0, \
-    "uilabel": "Geographical longitude in degrees as a double", \
-    "doc": "Longitude of the agent's geographical position. The value should " \
-           "be expressed in degrees as a double." \
-  }
-  double longitude;
-  // clang-format on
-
-  cyclus::toolkit::Position coordinates;
-
-  /// Records an agent's latitude and longitude to the output db
-  void RecordPosition();
+  private:
+  // Code Injection:
+  #include "toolkit/position.cycpp.h"
+  
 };
 
 }  // namespace cycamore