+++ src/sink.h
@@ -119,7 +119,10 @@ class Sink : public cyclus::Facility,
     return in_commod_prefs;
   }
 
-  private:
+ private:
+  // Code Injection:
+  #include "toolkit/position.cycpp.h"
+
   double requestAmt;
   int nextBuyTime;
 
@@ -286,26 +289,6 @@ class Sink : public cyclus::Facility,
   }
   bool keep_packaging;
 
-  #pragma cyclus var { \
-    "default": 0.0, \
-    "uilabel": "Geographical latitude in degrees as a double", \
-    "doc": "Latitude of the agent's geographical position. The value should "\
-           "be expressed in degrees as a double." \
-  }
-  double latitude;
-
-  #pragma cyclus var { \
-    "default": 0.0, \
-    "uilabel": "Geographical longitude in degrees as a double", \
-    "doc": "Longitude of the agent's geographical position. The value should "\
-           "be expressed in degrees as a double." \
-  }
-  double longitude;
-  // clang-format on
-
-  cyclus::toolkit::Position coordinates;
-
-  void RecordPosition();
 };
 
 }  // namespace cycamore